edit: swagger config
